rest of class list item component working
