+++ Frontend/app/class/browse/browse.html
@@ -6,6 +6,7 @@
     <p>
       These classes are available to all users and can be joined at any time. If you want to learn more about a given topic, please click the "Join <em>Class Name</em>" button that follows the description of the class. Note that classes you have already joined will not be shown in this list.
     </p>
+    <br/>
   </div>
 </div>
 <div class="row">
@@ -17,10 +18,7 @@
   <div class="col-xs-12" *ngIf="classes.length && !isLoading">
     <div class="row" *ngFor="let class of classes;">
       <div class="col-xs-12">
-        <class-list-item [classObject]="class"></class-list-item>
-        <button class="btn btn-sm btn-primary" (click)="joinClass(class.code)">
-          Join {{class.title}}
-        </button>
+        <class-list-item [classObject]="class" [joinClassButtonShown]="true"></class-list-item>
       </div>
     </div>
   </div>
