+++ Frontend/app/class/listItem/classListItem.component.ts
@@ -1,0 +1,39 @@
+import { Component, Input } from '@angular/core';
+import { UserService } from '../../user/user.service';
+import { ToastsManager } from 'ng2-toastr/ng2-toastr';
+import { Router } from '@angular/router';
+
+@Component({
+  moduleId: module.id,
+  selector: 'class-list-item',
+  templateUrl: 'classListItem.html'
+})
+
+export class ClassListItemComponent {
+  @Input() classObject;
+  @Input() classDetailsButton;
+  @Input() joinClassButtonShown;
+  isJoinClassInprogress = false;
+
+  constructor(public toastr: ToastsManager, public router: Router, public userService: UserService ){
+  }
+
+  joinClass(classCode) {
+    this.isJoinClassInprogress = true;
+    this.userService.joinClass(classCode)
+    .subscribe(
+      data => this.handleJoinSuccess(data.json().payload),
+      err => this.handleJoinError(err)
+    );
+  }
+
+  handleJoinSuccess(json) {
+    this.toastr.success('You have joined ' + json.title);
+    this.router.navigateByUrl('/class/' + json.code);
+  }
+
+  handleJoinError(err) {
+    this.isJoinClassInprogress = false;
+    this.toastr.error('Unable to join class');
+  }
+}
