+++ Frontend/app/class/listItem/classListItem.html
@@ -10,6 +10,9 @@
       {{classObject.description}}
     </p>
     <br/>
+    <button class="btn btn-sm btn-primary" (click)="joinClass(classObject.code)" *ngIf="joinClassButtonShown" [ladda]="isJoinClassInprogress">
+      Join {{classObject.title}}
+    </button>
     <a [routerLink]="['/class/', classObject.code]" class="btn btn-sm btn-info">View Class Detail Page</a>
   </div>
 </div>
